# perf(core): skip unnecessary stat call in `tgzUtils.convertToZip`