--- a/packages/yarnpkg-core/sources/worker-zip/Worker.ts
+++ b/packages/yarnpkg-core/sources/worker-zip/Worker.ts
@@ -1,4 +1,4 @@
-import {PortablePath, ZipFS}                    from '@yarnpkg/fslib';
+import {PortablePath, statUtils, ZipFS}         from '@yarnpkg/fslib';
 import {getLibzipPromise}                       from '@yarnpkg/libzip';
 import {parentPort}                             from 'worker_threads';
 
@@ -13,7 +13,7 @@ parentPort.on(`message`, async (data: ConvertToZipPayload) => {
   const {opts, tgz, tmpFile} = data;
   const {compressionLevel, ...bufferOpts} = opts;
 
-  const zipFs = new ZipFS(tmpFile, {create: true, libzip: await getLibzipPromise(), level: compressionLevel});
+  const zipFs = new ZipFS(tmpFile, {create: true, libzip: await getLibzipPromise(), level: compressionLevel, stats: statUtils.makeDefaultStats()});
 
   // Buffers sent through Node are turned into regular Uint8Arrays
   const tgzBuffer = Buffer.from(tgz.buffer, tgz.byteOffset, tgz.byteLength);