**What's the problem this PR addresses?**

`tgzUtils.convertToZip` tries to get the stats of a file that never exists.

Ref https://github.com/yarnpkg/berry/issues/4057

**How did you fix it?**

Pass the default stats to the `ZipFS` constructor to skip the stat call.

**Benchmark results**

On the gatsby benchmark this removes 1918 statx calls when fetching packages

```sh
YARN_IGNORE_PATH=1 hyperfine -w 1 --prepare="rm -rf ./.yarn-global"\
  "node ./master.cjs --mode skip-build"\
  "node ./skip-zip-stat.cjs --mode skip-build"
Benchmark 1: node ./master.cjs --mode skip-build
  Time (mean ± σ):      7.906 s ±  0.201 s    [User: 51.286 s, System: 3.319 s]
  Range (min … max):    7.700 s …  8.375 s    10 runs

Benchmark 2: node ./skip-zip-stat.cjs --mode skip-build
  Time (mean ± σ):      7.847 s ±  0.191 s    [User: 51.128 s, System: 3.252 s]
  Range (min … max):    7.646 s …  8.325 s    10 runs

Summary
  'node ./skip-zip-stat.cjs --mode skip-build' ran
    1.01 ± 0.04 times faster than 'node ./master.cjs --mode skip-build'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.